Move recursive_apply directly where it's used, having a small algorithms.py just for this is too convoluted
